Add files via upload
This is a collection of study notes for understanding the Payment Card
Industry Data Security Standard. It simplifies the key requirements of PCI
DSS and thus makes it rather easier for any individual to learn and apply
these standards.
The purpose of this book is to give you a clear and organized information
that will walk you through each of the 12 PCI DSS requirements. Each
section explains what the requirement means, why it is important, and how it
can be implemented in a real-world setting. I have also included additional
sections, such as the Secure Software Development Lifecycle (SSDLC) and
the Least Privilege Approach, to give you extra insights into related topics.
These are the notes collected and put together by Mohamed Makram, using
free and publicly available resources. I've done my best to provide you with
organized and simplified information that's accessible for everyone. Keep in
mind that this book is not an official PCI DSS guide; it's just a kind of
personal contribution to people who would like to learn about compliance with
the standard.
I hope this book will be helpful for you on your learning journey-be that
exam preparation, working on a compliance project, or simply to improve
your understanding of data security standards. Thank you for taking the time
to read these notes, and I wish you the best of luck with your studies in PCI
DSS!